+++ esther/libesther/src/common/bytearray.h
@@ -5,11 +5,11 @@
 namespace es {
 
 class ByteArray {
-    uint32_t size, capacity;
+    size_t size, capacity;
     uint8_t *data;
 
 public:
-    ByteArray(uint32_t initialCapacity = 0);
+    ByteArray(size_t initialCapacity = 0);
 
     ByteArray(const ByteArray &array);
     ByteArray(ByteArray &&array);
@@ -19,7 +19,7 @@ public:
     ByteArray &operator=(const ByteArray &array);
     ByteArray &operator=(ByteArray &&array);
 
-    uint8_t *allocate(uint32_t count);
+    uint8_t *allocate(size_t count);
     int reallocate();
 
     template <class T>
@@ -30,14 +30,14 @@ public:
 
     uint8_t &operator[](int index);
 
-    bool free(uint32_t count);
+    bool free(size_t count);
     void release();
 
-    bool enoughSpace(uint32_t count) const;
+    bool enoughSpace(size_t count) const;
 
     uint8_t *getData() const;
-    uint32_t getSize() const;
-    uint32_t getCapacity() const;
+    size_t getSize() const;
+    size_t getCapacity() const;
 };
 
 template <class T>
